First prototype.
It's fairly documented the main goal and the ideas.

I just put and got a ~113 MB file.  It was rather slow: 19s if the
chunks were not in the cluster, up to 67s if the chunks were being
overwritten (not the usual use case).

The test was done on two VMs with 256MB RAM.  Swapping was on the roof.

Increasing the RAM of both nodes to 512MB reduced the clean write to
8s-10s, and the overwrite to 12s-16s.  Swapping was reduced but still
present.

Both test used the protobuf client.

Setting up the bucket type's write_once property to True, inserting the
113 MB file under a new name took under 4s.  Rewriting (whatever the
semantics may be, but since the written data is always the same) took
about 8.4s.