Disallow overwriting a file.
The BlobWriter is not really designed to overwrite a blob.  If you overwrite a
large blob with a smaller one, the excess of chunks will be leaked in the
Riak KV store, and there's no easy way to reclaim the space.

So it's best to ensure you cannot overwrite.

This requires modifications in the FS tests, to ensure each test runs a clean
FS.

If the FS wants to support overwriting of files, it must ensure to first
remove the old blob.